remove markdown config
I can make it continue the list with enter and o but this then also
applies to textwidth wrapping which is what I do not want. I don't see a
formatoption to separate the two. As fo-c is formatting the comment and
inserting the comment leader. I do want the comment formatting but not
the insertion of the leader on formatting.